+++ libraries/Harmonical/Harmonical.h
@@ -43,7 +43,7 @@ class Harmonical {
   ~Harmonical();
 
   unsigned int GCD(unsigned int a, unsigned int b);	// greatest common divisor, Euklid
-  unsigned int LCM(unsigned int a, unsigned int b);	// least common multiple
+  unsigned long LCM(unsigned int a, unsigned int b);	// least common multiple
   void reduce_fraction(struct fraction *f);
   void expand_fractions(struct fraction * a, struct fraction * b);
   fraction * add_fraction(struct fraction * delta, struct fraction * sum);
